+++ frontend/catalyst/compiler.py
@@ -15,13 +15,15 @@
 MLIR/LLVM representations.
 """
 
+import abc
 import os
 import sys
 import shutil
 import subprocess
+import tempfile
 import warnings
 from io import TextIOWrapper
-from typing import Optional, List
+from typing import Optional, List, Any
 from dataclasses import dataclass
 
 from catalyst._configuration import INSTALLED
@@ -33,26 +35,22 @@ package_root = os.path.dirname(__file__)
 class CompileOptions:
     """Generic compilation options"""
 
-    verbose: bool
-    logfile: Optional[TextIOWrapper] = None  # stdout/stderr or a file
-
-    def get_logfile(self) -> TextIOWrapper:
-        """Get the effective file object, as configured"""
-        return self.logfile if self.logfile else sys.stderr
-
-
-default_compile_options: CompileOptions = CompileOptions(0, None)
+    verbose: Optional[bool] = False
+    logfile: Optional[TextIOWrapper] = sys.stderr
+    target: Optional[str] = "binary"
+    keep_intermediate: Optional[bool] = False
+    pipelines: Optional[List[Any]] = None
 
 
 def run_writing_command(
     command: List[str], compile_options: Optional[CompileOptions] = None
 ) -> None:
     """Run the command after optionally announcing this fact to the user"""
-    compile_options: CompileOptions = (
-        compile_options if compile_options else default_compile_options
-    )
+    if compile_options is None:
+        compile_options = CompileOptions()
+
     if compile_options.verbose:
-        print(f"[RUNNING] {' '.join(command)}", file=compile_options.get_logfile())
+        print(f"[RUNNING] {' '.join(command)}", file=compile_options.logfile)
     subprocess.run(command, check=True)
 
 
@@ -75,89 +73,199 @@ def get_executable_path(project, tool):
     return executable_path if os.path.exists(executable_path) else tool
 
 
-def get_lib_path(project, env):
+def get_lib_path(project, env_var):
     """Get the library path."""
-    return (
-        os.path.join(package_root, "lib")
-        if INSTALLED
-        else os.getenv(env, default_lib_paths.get(project, ""))
-    )
+    if INSTALLED:
+        return os.path.join(package_root, "lib")  # pragma: no cover
+    return os.getenv(env_var, default_lib_paths.get(project, ""))
 
 
-translate_tool = get_executable_path("llvm", "mlir-translate")
-mhlo_opt_tool = get_executable_path("mhlo", "mlir-hlo-opt")
-quantum_opt_tool = get_executable_path("quantum", "quantum-opt")
+class PassPipeline(abc.ABC):
+    """Abstract PassPipeline class."""
 
-mhlo_lowering_pass_pipeline = [
-    "--canonicalize",
-    "--chlo-legalize-to-hlo",
-    "--mhlo-legalize-control-flow",
-    "--hlo-legalize-to-linalg",
-    "--mhlo-legalize-to-std",
-    "--convert-to-signless",
-    "--canonicalize",
-]
+    _executable: Optional[str] = None
+    _default_flags: Optional[list[str]] = None
 
-quantum_compilation_pass_pipeline = [
-    "--lower-gradients",
-]
+    @staticmethod
+    @abc.abstractmethod
+    def get_output_filename(infile):
+        """Compute the output filename from the input filename.
 
-bufferization_pass_pipeline = [
-    "--inline",
-    "--gradient-bufferize",
-    "--scf-bufferize",
-    "--convert-tensor-to-linalg",  # tensor.pad
-    "--convert-elementwise-to-linalg",  # Must be run before --arith-bufferize
-    "--arith-bufferize",
-    "--empty-tensor-to-alloc-tensor",
-    "--bufferization-bufferize",
-    "--tensor-bufferize",
-    "--linalg-bufferize",
-    "--tensor-bufferize",
-    "--quantum-bufferize",
-    "--func-bufferize",
-    "--finalizing-bufferize",
-    "--buffer-hoisting",
-    "--buffer-loop-hoisting",
-    "--promote-buffers-to-stack",
-    "--buffer-deallocation",
-    "--convert-bufferization-to-memref",
-    "--canonicalize",
-    "--cse",
-]
+        .. note:
 
-llvm_lowering_pass_pipeline = [
-    "--convert-linalg-to-loops",
-    "--convert-scf-to-cf",
-    # This pass expands memref operations that modify the metadata of a memref (sizes, offsets,
-    # stdies) into a sequence of easier to analyze constructs. In particular, this pass transforms
-    # operations into explicit sequence of operations that model the effect of this operation on the
-    # different metadata. This pass uses affine constructs to materialize these effects.
-    # Concretely, expanded-strided-metadata is used to decompose memref.subview as it has no
-    # lowering in -convert-memref-to-llvm.
-    "--expand-strided-metadata",
-    "--lower-affine",
-    "--convert-complex-to-standard",  # added for complex.exp lowering
-    "--convert-complex-to-llvm",
-    "--convert-math-to-llvm",
-    # Must be run after -convert-math-to-llvm as it marks math::powf illegal but doesn't convert it.
-    "--convert-math-to-libm",
-    "--convert-arith-to-llvm",
-    "--convert-memref-to-llvm=use-generic-functions",
-    "--convert-index-to-llvm",
-    "--convert-gradient-to-llvm",
-    "--convert-quantum-to-llvm",
-    # Remove any dead casts as the final pass expects to remove all existing casts,
-    # but only those that form a loop back to the original type.
-    "--canonicalize",
-    "--reconcile-unrealized-casts",
-]
+                Derived classes are expected to implement this method.
 
-compiler = get_executable_path("llvm", "llc")
-compiler_flags = [
-    "--filetype=obj",
-    "--relocation-model=pic",
-]
+        Args:
+            infile (str): input file
+        Returns:
+            outfile (str): output file
+        """
+
+    @staticmethod
+    def _run(infile, outfile, executable, flags, options):
+        command = [executable] + flags + [infile, "-o", outfile]
+        run_writing_command(command, options)
+
+    @classmethod
+    # pylint: disable=too-many-arguments
+    def run(cls, infile, outfile=None, executable=None, flags=None, options=None):
+        """Run the pass.
+
+        Args:
+            infile (str): path to MLIR file to be compiled
+            outfile (str): path to output file, defaults to replacing extension in infile to .nohlo
+            executable (str): path to executable, defaults to mlir-hlo-opt
+            flags (List[str]): flags to mlir-hlo-opt, defaults to _default_flags
+            options (CompileOptions): compile options
+        """
+        if outfile is None:
+            outfile = cls.get_output_filename(infile)
+        if executable is None:
+            executable = cls._executable
+        if executable is None:
+            raise ValueError("Executable not specified.")
+        if flags is None:
+            flags = cls._default_flags
+        cls._run(infile, outfile, executable, flags, options)
+        return outfile
+
+
+# pylint: disable=too-few-public-methods
+class MHLOPass(PassPipeline):
+    """Pass pipeline to convert (M)HLO dialects to standard MLIR dialects."""
+
+    _executable = get_executable_path("mhlo", "mlir-hlo-opt")
+    _default_flags = [
+        "--allow-unregistered-dialect",
+        "--canonicalize",
+        "--chlo-legalize-to-hlo",
+        "--mhlo-legalize-control-flow",
+        "--hlo-legalize-to-linalg",
+        "--mhlo-legalize-to-std",
+        "--convert-to-signless",
+        "--canonicalize",
+    ]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".mlir"):
+            raise ValueError(f"Input file ({infile}) for MHLO is not an MLIR file")
+        return infile.replace(".mlir", ".nohlo.mlir")
+
+
+class BufferizationPass(PassPipeline):
+    """Pass pipeline that bufferizes MLIR dialects."""
+
+    _executable = get_executable_path("quantum", "quantum-opt")
+    _default_flags = [
+        "--inline",
+        "--gradient-bufferize",
+        "--scf-bufferize",
+        "--convert-tensor-to-linalg",  # tensor.pad
+        "--convert-elementwise-to-linalg",  # Must be run before --arith-bufferize
+        "--arith-bufferize",
+        "--empty-tensor-to-alloc-tensor",
+        "--bufferization-bufferize",
+        "--tensor-bufferize",
+        "--linalg-bufferize",
+        "--tensor-bufferize",
+        "--quantum-bufferize",
+        "--func-bufferize",
+        "--finalizing-bufferize",
+        "--buffer-hoisting",
+        "--buffer-loop-hoisting",
+        "--buffer-deallocation",
+        "--convert-bufferization-to-memref",
+        "--canonicalize",
+        "--cse",
+    ]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".opt.mlir"):
+            raise ValueError(f"Input file ({infile}) for bufferization is not an MLIR file")
+
+        return infile.replace(".opt.mlir", ".buff.mlir")
+
+
+class MLIRToLLVMDialect(PassPipeline):
+    """Pass pipeline to lower MLIR dialects to LLVM dialect."""
+
+    _executable = get_executable_path("quantum", "quantum-opt")
+    _default_flags = [
+        "--convert-linalg-to-loops",
+        "--convert-scf-to-cf",
+        # This pass expands memref operations that modify the metadata of a memref (sizes, offsets,
+        # stdies) into a sequence of easier to analyze constructs. In particular, this pass transforms
+        # operations into explicit sequence of operations that model the effect of this operation on the
+        # different metadata. This pass uses affine constructs to materialize these effects.
+        # Concretely, expanded-strided-metadata is used to decompose memref.subview as it has no
+        # lowering in -convert-memref-to-llvm.
+        "--expand-strided-metadata",
+        "--lower-affine",
+        "--convert-complex-to-standard",  # added for complex.exp lowering
+        "--convert-complex-to-llvm",
+        "--convert-math-to-llvm",
+        # Must be run after -convert-math-to-llvm as it marks math::powf illegal but doesn't convert it.
+        "--convert-math-to-libm",
+        "--convert-arith-to-llvm",
+        "--convert-memref-to-llvm",
+        "--convert-index-to-llvm",
+        "--convert-gradient-to-llvm",
+        "--convert-quantum-to-llvm",
+        # Remove any dead casts as the final pass expects to remove all existing casts,
+        # but only those that form a loop back to the original type.
+        "--canonicalize",
+        "--reconcile-unrealized-casts",
+    ]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".buff.mlir"):
+            raise ValueError(f"Input file ({infile}) is not a bufferized MLIR file")
+        return infile.replace(".buff.mlir", ".llvm.mlir")
+
+
+class QuantumCompilationPass(PassPipeline):
+    """Pass pipeline to lower gradients."""
+
+    _executable = get_executable_path("quantum", "quantum-opt")
+    _default_flags = ["--lower-gradients"]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".mlir"):
+            raise ValueError(f"Input file ({infile}) for quantum transforms is not an MLIR file")
+        return infile.replace(".mlir", ".opt.mlir")
+
+
+class LLVMDialectToLLVMIR(PassPipeline):
+    """Convert LLVM Dialect to LLVM-IR."""
+
+    _executable = get_executable_path("llvm", "mlir-translate")
+    _default_flags = ["--mlir-to-llvmir"]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".llvm.mlir"):
+            raise ValueError(f"Input file ({infile}) is not an LLVM dialect MLIR file")
+        return infile.replace(".llvm.mlir", ".ll")
+
+
+class LLVMIRToObjectFile(PassPipeline):
+    """LLVMIR To Object File."""
+
+    _executable = get_executable_path("llvm", "llc")
+    _default_flags = [
+        "--filetype=obj",
+        "--relocation-model=pic",
+    ]
+
+    @staticmethod
+    def get_output_filename(infile):
+        if not infile.endswith(".ll"):
+            raise ValueError(f"Input file ({infile}) for compilation is not an LLVMIR file")
+        return infile.replace(".ll", ".o")
 
 
 # pylint: disable=too-few-public-methods
@@ -178,31 +286,34 @@ class CompilerDriver:
     _default_fallback_compilers = ["clang", "gcc", "c99", "c89", "cc"]
 
     @staticmethod
-    def _flags():
-        mlir_lib_path = get_lib_path("llvm", "MLIR_LIB_DIR")
-        lrt_lib_path = get_lib_path("runtime", "RUNTIME_LIB_DIR")
-        lrt_capi_path = os.path.join(lrt_lib_path, "capi")
-        lrt_backend_path = os.path.join(lrt_lib_path, "backend")
+    def get_default_flags():
+        """Re-compute the path where the libraries exist.
 
-        flags = [
+        The use case for this is if someone is in a python jupyter notebook and
+        needs to change the environment mid computation.
+        Returns
+            (List[str]): The default flag list.
+        """
+        mlir_lib_path = get_lib_path("llvm", "MLIR_LIB_DIR")
+        rt_lib_path = get_lib_path("runtime", "RUNTIME_LIB_DIR")
+        rt_capi_path = os.path.join(rt_lib_path, "capi")
+        rt_backend_path = os.path.join(rt_lib_path, "backend")
+
+        default_flags = [
             "-shared",
             "-rdynamic",
-            f"-L{mlir_lib_path}",
             "-Wl,-no-as-needed",
-            f"-Wl,-rpath,{mlir_lib_path}",
-            f"-L{lrt_capi_path}",
-            f"-L{lrt_backend_path}",
-            f"-Wl,-rpath,{lrt_capi_path}:{lrt_backend_path}",
-            f"-L{lrt_capi_path}",
-            f"-L{lrt_backend_path}",
-            f"-Wl,-rpath,{lrt_capi_path}:{lrt_backend_path}",
+            f"-Wl,-rpath,{rt_capi_path}:{rt_backend_path}:{mlir_lib_path}",
+            f"-L{mlir_lib_path}",
+            f"-L{rt_capi_path}",
+            f"-L{rt_backend_path}",
             "-lrt_backend",
             "-lrt_capi",
             "-lpthread",
             "-lmlir_c_runner_utils",  # required for memref.copy
         ]
 
-        return flags
+        return default_flags
 
     @staticmethod
     def _get_compiler_fallback_order(fallback_compilers):
@@ -234,13 +345,10 @@ class CompilerDriver:
 
     @staticmethod
     # pylint: disable=redefined-outer-name
-    def _attempt_link(compiler, flags, infile, outfile, compile_options=None):
-        compile_options = compile_options if compile_options else default_compile_options
+    def _attempt_link(compiler, flags, infile, outfile, options):
         try:
             command = [compiler] + flags + [infile, "-o", outfile]
-            if compile_options.verbose:
-                print(f"[RUNNING] {' '.join(command)}", file=compile_options.get_logfile())
-            subprocess.run(command, check=True)
+            run_writing_command(command, options)
             return True
         except subprocess.CalledProcessError:
             msg = (
@@ -251,7 +359,19 @@ class CompilerDriver:
             return False
 
     @staticmethod
-    def link(infile, outfile, flags=None, fallback_compilers=None, compile_options=None):
+    def get_output_filename(infile):
+        """Rename object file to shared object
+
+        Args:
+            infile (str): input file name
+            outfile (str): output file name
+        """
+        if not infile.endswith(".o"):
+            raise ValueError(f"Input file ({infile}) is not an object file")
+        return infile.replace(".o", ".so")
+
+    @staticmethod
+    def run(infile, outfile=None, flags=None, fallback_compilers=None, options=None):
         """
         Link the infile against the necessary libraries and produce the outfile.
 
@@ -264,233 +384,111 @@ class CompilerDriver:
         Raises:
             EnvironmentError: The exception is raised when no compiler succeeded.
         """
+        if outfile is None:
+            outfile = CompilerDriver.get_output_filename(infile)
         if flags is None:
-            flags = CompilerDriver._flags()
+            flags = CompilerDriver.get_default_flags()
         if fallback_compilers is None:
             fallback_compilers = CompilerDriver._default_fallback_compilers
-        # pylint: disable=redefined-outer-name
         for compiler in CompilerDriver._available_compilers(fallback_compilers):
-            success = CompilerDriver._attempt_link(
-                compiler, flags, infile, outfile, compile_options
-            )
+            success = CompilerDriver._attempt_link(compiler, flags, infile, outfile, options)
             if success:
-                return
+                return outfile
         msg = f"Unable to link {infile}. All available compiler options exhausted. Please provide a compatible compiler via $CATALYST_CC."
         raise EnvironmentError(msg)
 
 
-def lower_mhlo_to_linalg(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Translate MHLO to linalg dialect.
+class Compiler:
+    """Compiles MLIR modules to shared objects."""
 
-    Args:
-        filename (str): the path to a file were the program is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-5:] != ".mlir":
-        raise ValueError(f"Input file ({filename}) for MHLO lowering is not an MLIR file")
+    def __init__(self):
+        self.pass_pipeline_output = None
+        # The temporary directory must be referenced by the wrapper class
+        # in order to avoid being garbage collected
+        # pylint: disable=consider-using-with
+        self.workspace = tempfile.TemporaryDirectory()
 
-    new_fname = filename.replace(".mlir", ".nohlo.mlir")
+    def run(self, mlir_module, options):
+        """Compile an MLIR module to a shared object.
 
-    command = [mhlo_opt_tool]
-    command += ["--allow-unregistered-dialect"]
-    command += [filename]
-    command += mhlo_lowering_pass_pipeline
-    command += ["-o", new_fname]
+        .. note::
 
-    run_writing_command(command, compile_options)
+            For compilation of hybrid quantum-classical PennyLane programs,
+            please see the :func:`~.qjit` decorator.
 
-    return new_fname
+        Args:
+            compile_options (Optional[CompileOptions]): common compilation options
 
+        Returns:
+            (str): filename of shared object
+        """
 
-def transform_quantum_ir(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Runs quantum optimizations and transformations, as well gradient transforms, on the hybrid
-    IR.
+        module_name = mlir_module.operation.attributes["sym_name"]
+        # Convert MLIR string to Python string
+        module_name = str(module_name)
+        # Remove quotations
+        module_name = module_name.replace('"', "")
 
-    Args:
-        filename (str): the path to a file where the program is stored.
-    Returns:
-        a path to the output file
-    """
-    if filename[-5:] != ".mlir":
-        raise ValueError(f"Input file ({filename}) for quantum transforms is not an MLIR file")
+        if options.keep_intermediate:
+            parent_dir = os.getcwd()
+            path = os.path.join(parent_dir, module_name)
+            os.makedirs(path, exist_ok=True)
+            workspace_name = os.path.abspath(path)
+        else:
+            workspace_name = self.workspace.name
 
-    new_fname = filename.replace(".mlir", ".opt.mlir")
+        pipelines = options.pipelines
+        if pipelines is None:
+            pipelines = [
+                MHLOPass,
+                QuantumCompilationPass,
+                BufferizationPass,
+                MLIRToLLVMDialect,
+                LLVMDialectToLLVMIR,
+                LLVMIRToObjectFile,
+                CompilerDriver,
+            ]
 
-    command = [quantum_opt_tool]
-    command += [filename]
-    command += quantum_compilation_pass_pipeline
-    command += ["-o", new_fname]
+        self.pass_pipeline_output = {}
 
-    run_writing_command(command, compile_options)
+        filename = f"{workspace_name}/{module_name}.mlir"
+        with open(filename, "w", encoding="utf-8") as f:
+            mlir_module.operation.print(f, print_generic_op_form=False, assume_verified=True)
 
-    return new_fname
+        for pipeline in pipelines:
+            output = pipeline.run(filename, options=options)
+            self.pass_pipeline_output[pipeline] = output
+            filename = os.path.abspath(output)
 
+        return filename
 
-def bufferize_tensors(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Translate MHLO to linalg dialect.
+    @staticmethod
+    def _get_class_from_string(pipeline):
+        return getattr(sys.modules[__name__], pipeline)
 
-    Args:
-        filename (str): the path to a file were the program is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-5:] != ".mlir":
-        raise ValueError(f"Input file ({filename}) for bufferization is not an MLIR file")
+    def _get_output_file_of(self, pipeline):
+        cls = Compiler._get_class_from_string(pipeline)
+        return self.pass_pipeline_output.get(cls)
 
-    new_fname = filename.replace(".mlir", ".buff.mlir")
+    def get_output_of(self, pipeline):
+        """Get the output IR of a pipeline.
+        Args:
+            pipeline (str): name of pass class
 
-    command = [quantum_opt_tool]
-    command += [filename]
-    command += bufferization_pass_pipeline
-    command += ["-o", new_fname]
+        Returns
+            (str): output IR
+        """
+        try:
+            fname = self._get_output_file_of(pipeline)
+        except AttributeError as e:
+            raise ValueError(f"Output for pass {pipeline} not found.") from e
+        with open(fname, "r", encoding="utf-8") as f:
+            txt = f.read()
+        return txt
 
-    run_writing_command(command, compile_options)
-
-    return new_fname
-
-
-def lower_all_to_llvm(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Translate MLIR dialects to LLVM dialect.
-
-    Args:
-        filename (str): the path to a file were the program is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-10:] != ".buff.mlir":
-        raise ValueError(f"Input file ({filename}) for LLVM lowering is not a bufferized MLIR file")
-
-    new_fname = filename.replace(".buff.mlir", ".llvm.mlir")
-
-    command = [quantum_opt_tool]
-    command += [filename]
-    command += llvm_lowering_pass_pipeline
-    command += ["-o", new_fname]
-
-    run_writing_command(command, compile_options)
-
-    return new_fname
-
-
-def convert_mlir_to_llvmir(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Translate LLVM dialect to LLVM IR.
-
-    Args:
-        filename (str): the path to a file were the program is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-10:] != ".llvm.mlir":
-        raise ValueError(
-            f"Input file ({filename}) for LLVMIR conversion is not an LLVM dialect MLIR file"
-        )
-
-    new_fname = filename.replace(".llvm.mlir", ".ll")
-
-    command = [translate_tool]
-    command += [filename]
-    command += ["--mlir-to-llvmir"]
-    command += ["-o", new_fname]
-
-    run_writing_command(command, compile_options)
-
-    return new_fname
-
-
-def compile_llvmir(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Translate LLVM IR to an object file.
-
-    Args:
-        filename (str): the path to a file were the program is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-3:] != ".ll":
-        raise ValueError(f"Input file ({filename}) for compilation is not an LLVMIR file")
-
-    new_fname = filename.replace(".ll", ".o")
-
-    command = [compiler]
-    command += compiler_flags
-    command += [filename]
-    command += ["-o", new_fname]
-
-    run_writing_command(command, compile_options)
-
-    return new_fname
-
-
-def link_lightning_runtime(filename: str, compile_options: Optional[CompileOptions] = None) -> str:
-    """Link the object file as a shared object.
-
-    Args:
-        filename (str): the path to a file were the object file is stored.
-        Optional compile_options (CompileOptions): generic compilation options.
-    Returns:
-        a path to the output file
-    """
-    if filename[-2:] != ".o":
-        raise ValueError(f"Input file ({filename}) for linking is not an object file")
-
-    new_fname = filename.replace(".o", ".so")
-
-    CompilerDriver.link(filename, new_fname, compile_options=compile_options)
-
-    return new_fname
-
-
-def compile(mlir_module, workspace, passes, compile_options: Optional[CompileOptions] = None):
-    """Compile an MLIR module to a shared object.
-
-    .. note::
-
-        For compilation of hybrid quantum-classical PennyLane programs,
-        please see the :func:`~.qjit` decorator.
-
-    Args:
-        mlir_module (Module): the MLIR module
-        workspace (str): the absolute path to the MLIR module
-        has_hlo (bool): ``True`` if the MLIR module contains HLO code. Defaults to ``False``
-        passes (List[str]): the list of compilation passes
-        Optional compile_options (CompileOptions): generic compilation options.
-
-    Returns:
-        Shared object
-        A string representation of LLVM IR.
-    """
-
-    module_name = mlir_module.operation.attributes["sym_name"]
-    # Convert MLIR string to Python string
-    module_name = str(module_name)
-    # Remove quotations
-    module_name = module_name.replace('"', "")
-    # need to create a temporary file with the string contents
-    filename = f"{workspace}/{module_name}.mlir"
-    with open(filename, "w", encoding="utf-8") as f:
-        mlir_module.operation.print(f, print_generic_op_form=False, assume_verified=True)
-
-    mlir = filename
-    passes["mlir"] = mlir
-    nohlo = lower_mhlo_to_linalg(mlir, compile_options)
-    passes["nohlo"] = nohlo
-    optimized = transform_quantum_ir(nohlo, compile_options)
-    passes["opt"] = optimized
-    buff = bufferize_tensors(optimized, compile_options)
-    passes["buff"] = buff
-    llvm_dialect = lower_all_to_llvm(buff, compile_options)
-    passes["llvm"] = llvm_dialect
-    llvmir = convert_mlir_to_llvmir(llvm_dialect, compile_options)
-    passes["ll"] = llvmir
-    object_file = compile_llvmir(llvmir, compile_options)
-    shared_object = link_lightning_runtime(object_file, compile_options)
-
-    with open(llvmir, "r", encoding="utf-8") as f:
-        _llvmir = f.read()
-
-    return shared_object, _llvmir
+    def print(self, pipeline):
+        """Print the output IR of pass.
+        Args:
+            pipeline (str): name of pass class
+        """
+        print(self.get_output_of(pipeline))  # pragma: no cover